Adding new code by Greg Haerr: a 64bit version of the MWin API, X11 driver enhancements, Apple OS x support and demo code for using the agg lite library with nano-X and the canny edge detector.

+++ src/include/windows.h
@@ -146,8 +146,8 @@ struct hwnd {
 	HBRUSH		paintBrush;	/* brush created to paint some controls */
 	HPEN		paintPen;	/* pen created to paint some controls */
 	MWCLIPREGION *	update;		/* update region in screen coords*/
-	DWORD		userdata;	/* setwindowlong user data*/
-	DWORD		userdata2;	/* additional user data (will remove)*/
+	LONG		userdata;	/* setwindowlong user data*/
+	LONG		userdata2;	/* additional user data (will remove)*/
 	MWLISTHEAD  	props;		/* windows property list */
 	MWSCROLLBARINFO	hscroll;	/* NC scrollbars*/
 	MWSCROLLBARINFO	vscroll;
